move allocate_point to sierra_utils.rs
