+++ crates/blockifier/src/execution/sierra_utils.rs
@@ -2,7 +2,8 @@ use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
 use std::rc::Rc;
 
-use ark_ff::BigInt;
+use ark_ec::short_weierstrass::SWCurveConfig;
+use ark_ff::{BigInt, PrimeField};
 use cairo_lang_sierra::ids::FunctionId;
 use cairo_lang_sierra::program::Program as SierraProgram;
 use cairo_lang_starknet_classes::contract_class::{ContractEntryPoint, ContractEntryPoints};
@@ -28,7 +29,8 @@ use super::entry_point::{CallEntryPoint, EntryPointExecutionResult};
 use super::errors::EntryPointExecutionError;
 use super::native_syscall_handler::NativeSyscallHandler;
 use crate::execution::entry_point::EntryPointExecutionContext;
-use crate::execution::syscalls::secp::SecpHintProcessor;
+use crate::execution::syscalls::hint_processor::SyscallExecutionError;
+use crate::execution::syscalls::secp::{SecpHintProcessor, SecpNewRequest, SecpNewResponse};
 use crate::state::state_api::State;
 
 // An arbitrary number, chosen to avoid accidentally aligning with actually calculated gas
@@ -289,4 +291,28 @@ pub fn decode_felts_as_str(encoding: &[Felt]) -> String {
         encoding.iter().flat_map(|felt| felt.to_bytes_be()[1..32].to_vec()).collect();
 
     String::from_utf8(bytes_err).unwrap().trim_matches('\0').to_owned()
+}
+
+pub fn allocate_point<Curve: SWCurveConfig>(
+    point_x: U256,
+    point_y: U256,
+    hint_processor: &mut SecpHintProcessor<Curve>,
+) -> cairo_native::starknet::SyscallResult<usize>
+where
+    Curve::BaseField: PrimeField,
+{
+    let request = SecpNewRequest { x: u256_to_biguint(point_x), y: u256_to_biguint(point_y) };
+
+    let response = hint_processor.secp_new_unchecked(request);
+
+    match response {
+        // We can't receive None here, as the response is always Some from `secp_new_unchecked`.
+        Ok(SecpNewResponse { optional_ec_point_id: id }) => Ok(id.unwrap()),
+        Err(SyscallExecutionError::SyscallError { error_data }) => {
+            Err(error_data.iter().map(|felt| starkfelt_to_felt(*felt)).collect())
+        }
+        Err(_) => unreachable!(
+            "Can't receive an error other than SyscallError from `secp_new_unchecked`."
+        ),
+    }
 }
