+++ crates/blockifier/src/execution/syscalls/hint_processor.rs
@@ -168,18 +168,6 @@ pub const FAILED_TO_EXECUTE_CALL: &str = "0x004661696c656420746f2065786563757465
 // Failed to calculate address
 pub const FAILED_TO_CALCULATE_CONTRACT_ADDRESS: &str =
     "0x004661696c656420746f2063616c63756c6174652061646472657373";
-// Failed to parse
-pub const FAILED_TO_PARSE: &str = "0x004661696c656420746f20706172736520";
-// Failed to read result
-pub const FAILED_TO_READ_RESULT: &str = "0x004661696c656420746f207265616420726573756c7420";
-// Failed to write
-pub const FAILED_TO_WRITE: &str = "0x004661696c656420746f20777269746520";
-// Invalid point
-pub const INVALID_POINT: &str = "0x00496e76616c696420706f696e74";
-// Unreachable error
-pub const UNREACHABLE_ERROR: &str = "0x00556e726561636861626c65206572726f72";
-// Invalid scalar
-pub const INVALID_SCALAR: &str = "0x00496e76616c6964207363616c6172";
 
 #[cfg(test)]
 mod tests {
@@ -196,12 +184,6 @@ mod tests {
         assert!(Felt::from_hex(L2_GAS).is_ok());
         assert!(Felt::from_hex(FAILED_TO_EXECUTE_CALL).is_ok());
         assert!(Felt::from_hex(FAILED_TO_CALCULATE_CONTRACT_ADDRESS).is_ok());
-        assert!(Felt::from_hex(FAILED_TO_PARSE).is_ok());
-        assert!(Felt::from_hex(FAILED_TO_READ_RESULT).is_ok());
-        assert!(Felt::from_hex(FAILED_TO_WRITE).is_ok());
-        assert!(Felt::from_hex(INVALID_POINT).is_ok());
-        assert!(Felt::from_hex(UNREACHABLE_ERROR).is_ok());
-        assert!(Felt::from_hex(INVALID_SCALAR).is_ok());
     }
 }
